requirements.txt
requests===2.25.1
setuptools===56.0.0
python-dotenv===0.17.0
pandas===1.2.4